# PERF: HDFStore __unicode__ method