--- a/ci/requirements-2.7.build
+++ b/ci/requirements-2.7.build
@@ -2,5 +2,5 @@ python=2.7*
 python-dateutil=2.4.1
 pytz=2013b
 nomkl
-numpy
+numpy=1.12*
 cython=0.23