--- a/ci/requirements-3.6_DOC.build
+++ b/ci/requirements-3.6_DOC.build
@@ -1,5 +1,5 @@
 python=3.6*
 python-dateutil
 pytz
-numpy
+numpy=1.12*
 cython