--- a/ci/requirements-3.6_DOC.build
+++ b/ci/requirements-3.6_DOC.build
@@ -1,4 +1,4 @@
-python=3.5*
+python=3.6*
 python-dateutil
 pytz
 numpy