--- a/ci/requirements-3.6_DOC.run
+++ b/ci/requirements-3.6_DOC.run
@@ -12,7 +12,7 @@ lxml
 beautifulsoup4
 html5lib
 pytables
-openpyxl=1.8.5
+openpyxl
 xlrd
 xlwt
 xlsxwriter
@@ -21,4 +21,4 @@ numexpr
 bottleneck
 statsmodels
 xarray
-pyqt=4.11.4
+pyqt