--- a/pandas/compat/numpy/__init__.py
+++ b/pandas/compat/numpy/__init__.py
@@ -15,6 +15,7 @@ _np_version_under1p10 = _nlv < '1.10'
 _np_version_under1p11 = _nlv < '1.11'
 _np_version_under1p12 = _nlv < '1.12'
 _np_version_under1p13 = _nlv < '1.13'
+_np_version_under1p14 = _nlv < '1.14'
 
 if _nlv < '1.7.0':
     raise ImportError('this version of pandas is incompatible with '
@@ -74,4 +75,6 @@ __all__ = ['np',
            '_np_version_under1p10',
            '_np_version_under1p11',
            '_np_version_under1p12',
+           '_np_version_under1p13',
+           '_np_version_under1p14'
            ]