… .travis.yml
… AUTHORS.md
… asv_bench/benchmarks/hdfstore_bench.py
… asv_bench/benchmarks/indexing.py
… ci/requirements-3.6_DOC.sh
… doc/source/advanced.rst
… doc/source/api.rst
… doc/source/groupby.rst
… doc/source/io.rst
… doc/source/release.rst
… doc/source/reshaping.rst
… doc/source/timeseries.rst
… doc/source/whatsnew.rst
… doc/source/whatsnew/v0.20.0.txt
… doc/source/whatsnew/v0.20.2.txt
… doc/source/whatsnew/v0.20.3.txt
… doc/source/whatsnew/v0.21.0.txt
… pandas/core/indexes/period.py
… pandas/core/missing.py
… pandas/core/reshape/concat.py
… pandas/core/reshape/pivot.py
… pandas/core/reshape/reshape.py
… pandas/core/reshape/tile.py
… pandas/core/sparse/array.py
… pandas/core/window.py
… pandas/io/common.py
… pandas/io/excel.py
… pandas/io/formats/format.py
… pandas/io/formats/terminal.py
… pandas/io/parsers.py