@jreback 2017-05-26
ping when fixed / green.