@gfyoung 2017-05-28
1) Got a small linting error in `pytables.py` (hence the Travis failure)

2) Can you provide some timing info in the PR given that this is a performance change?