@Kiv 2017-05-28
@jreback I think this is ready to go?