@gfyoung 2017-05-29
@Kiv Some timing info in the PR would be useful for future reference.  Plus a benchmark test would be good here.