@Kiv 2017-06-11
My git branch got out of date and I don't know how to update it properly. New version with ASV bench is here:

https://github.com/pandas-dev/pandas/pull/16666