@jreback 2017-06-11
in the future see: http://pandas.pydata.org/pandas-docs/stable/contributing.html#combining-commits

you just needed to do:

```
git rebase -i upstream/master
git push yourremove yourbranch -f
```